Add tracking weight feature

+++ src/main/java/com/example/fitjourneyapp/service/AuthService.java
@@ -6,7 +6,8 @@ import java.util.Date;
 
 public interface AuthService {
     User login(String username, String password);
-    User register(String username, String password, String repeatPassword, String name, String surname, Date dateOfBirth,double weight);
+    User register(String username, String password, String repeatPassword, String name, String surname, Date dateOfBirth,float weight);
     User save(User user);
-    User edit(String username, String password, String repeatPassword, String name, String surname, Date dateOfBirth, double weight);
+    User edit(String username, String password, String repeatPassword, String name, String surname, Date dateOfBirth, float weight);
+
 }
