+++ src/main/java/com/example/fitjourneyapp/web/controller/RegisterController.java
@@ -34,7 +34,7 @@ public class RegisterController {
     }
 
     @PostMapping
-    public String register(@RequestParam String username, @RequestParam String password, @RequestParam String repeatPassword, @RequestParam String name, @RequestParam String surname, @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) Date dateOfBirth, @RequestParam double weight)
+    public String register(@RequestParam String username, @RequestParam String password, @RequestParam String repeatPassword, @RequestParam String name, @RequestParam String surname, @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) Date dateOfBirth, @RequestParam float weight)
     {
         try {
             this.authService.register(username, password, repeatPassword, name, surname, dateOfBirth, weight);
